structure: move headers from include/ to include/compio
