+++ benchmarks/custom/fsize_benchmark.cpp
@@ -1,5 +1,5 @@
 #include "compio.h"
-#include "compio_file.hpp"
+#include "compio/compio_file.hpp"
 
 #include "sample_data.hpp"
