+++ include/compio/utils.hpp
@@ -7,7 +7,7 @@
 #define UTILS_HEADER_
 
 #include "compio.h"
-#include "compio_file.hpp"
+#include "compio/compio_file.hpp"
 
 namespace compio {
