+++ src/allocator.cpp
@@ -4,15 +4,15 @@
  */
 
 #include <cinttypes>
-#include "compio_file.hpp"
-#include "debug_print.hpp"
-#include "allocator.hpp"
-#include "file.hpp"
+#include "compio/compio_file.hpp"
+#include "compio/debug_print.hpp"
+#include "compio/allocator.hpp"
+#include "compio/file.hpp"
+#include "compio/utils.hpp"
 #include <algorithm>
 #include <cstdio>
 #include <iostream>
 #include <algorithm>
-#include <utils.hpp>
 #include <cassert>
 #include <cstring>
 #include <vector>
