+++ src/storage_block_reader.cpp
@@ -1,6 +1,6 @@
-#include "storage_block_reader.hpp"
+#include "compio/storage_block_reader.hpp"
 
-#include "debug_print.hpp"
+#include "compio/debug_print.hpp"
 
 namespace compio {
