+++ tests/src/test_allocator.cpp
@@ -1,6 +1,6 @@
-#include "allocator.hpp"
-#include "compio_file.hpp"
-#include "utils.hpp"
+#include "compio/allocator.hpp"
+#include "compio/compio_file.hpp"
+#include "compio/utils.hpp"
 #include <gtest/gtest.h>
 #include "test_util.hpp"
