+++ tests/src/test_library.cpp
@@ -3,7 +3,7 @@
 #include <random>
 
 #include "compio.h"
-#include "compio_file.hpp"
+#include "compio/compio_file.hpp"
 #include "sample_data.hpp"
 #include "test_util.hpp"
