+++ util/unpack.cpp
@@ -1,6 +1,6 @@
 #include "compio.h"
 
-#include "compio_file.hpp"
+#include "compio/compio_file.hpp"
 
 #include <stdexcept>
